Remove Event component, Added Features
