--- src/components/EventCreationPage.css
@@ -1,24 +1,0 @@
-
-.events-container {
-  padding: 2rem;
-  text-align: center;
-}
-
-.header-title h1 {
-  margin-bottom: 1.5rem;
-}
-
-.option-button {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 0.5rem;
-  margin: 1rem 0;
-}
-
-.error {
-  color: red;
-  font-size: 0.9rem;
-  margin-top: 0.5rem;
-}
-
