+++ src/pages/Events.css
@@ -1,0 +1,69 @@
+.events-container {
+    padding: 2rem;
+    text-align: center;
+  }
+  
+  .header-title h1 {
+    margin-bottom: 1.5rem;
+  }
+  
+  .option-button {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    gap: 0.5rem;
+    margin: 1rem 0;
+  }
+  
+  .error {
+    color: var(--RED);
+    font-size: 0.9rem;
+    margin-top: 0.5rem;
+  }
+
+  .event-list {
+    margin-top: 2rem;
+    padding: 1.5rem;
+    background-color: #ffffff;
+    border-radius: 1rem;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.08);
+  }
+  
+  .event-list h2 {
+    margin-bottom: 1rem;
+    text-align: center;
+  }
+  
+  .event-list-items {
+    list-style-type: none;
+    padding: 0;
+  }
+  
+  .event-card {
+    background-color: #f3f3f3;
+    margin-bottom: 1rem;
+    padding: 1rem;
+    border-radius: 0.75rem;
+    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.05);
+  }
+
+  .option-button {
+    display: flex;
+    gap: 1rem;
+    flex-wrap: wrap;
+    margin-top: 1rem;
+    justify-content: center;
+  }
+  
+  .event-card {
+    cursor: pointer;
+    transition: transform 0.15s ease;
+  }
+  
+  .event-card:hover {
+    transform: scale(1.01);
+  }
+  
+  
+  
+
